+++ client/src/hook/useContextOpv.js
@@ -1,0 +1,7 @@
+import { useContext } from 'react';
+import { OpenViduContext } from '../context/openViduContext';
+
+
+export const useContextOpv = () => {
+    return useContext(OpenViduContext);
+};
